moved userManagement state from root to ui state
