+++ src/client/components/EditUserForm/CountryRoles/CountryRoles.tsx
@@ -10,8 +10,8 @@ import { RoleName, User, UserRole, Users } from '@meta/user'
 import { useAppDispatch } from '@client/store'
 import { useCountries } from '@client/store/assessment'
 import { useAssessment, useCycle, useSecondaryRegion } from '@client/store/assessment/hooks'
+import { UserManagementActions } from '@client/store/ui/userManagement'
 import { useUser } from '@client/store/user'
-import { UserManagementActions } from '@client/store/userManagement'
 import CountrySelectModal from '@client/components/CountrySelectModal'
 
 import CountryRole from './CountryRole'
