+++ src/client/components/InviteUserForm/InviteUserForm.tsx
@@ -6,8 +6,8 @@ import { RoleName, Users } from '@meta/user'
 
 import { useAppDispatch } from '@client/store'
 import { useAssessment, useCycle } from '@client/store/assessment'
+import { UserManagementActions } from '@client/store/ui/userManagement'
 import { useUser } from '@client/store/user'
-import { UserManagementActions } from '@client/store/userManagement'
 import { useCountryIso } from '@client/hooks'
 import { useToaster } from '@client/hooks/useToaster'
