+++ src/client/components/UserList/ButtonUserListExport/ButtonUserListExport.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 
 import { RoleName, User, UserRole, Users } from '@meta/user'
 
-import { useUsers } from '@client/store/userManagement/hooks'
+import { useUsers } from '@client/store/ui/userManagement/hooks'
 import { useCountryIso } from '@client/hooks'
 import Icon from '@client/components/Icon'
