+++ src/client/components/UserList/UserList.tsx
@@ -11,8 +11,8 @@ import { User, Users, UserStatus } from '@meta/user'
 
 import { useAppDispatch } from '@client/store'
 import { useAssessment, useCycle } from '@client/store/assessment'
+import { UserManagementActions } from '@client/store/ui/userManagement'
 import { useUser } from '@client/store/user'
-import { UserManagementActions } from '@client/store/userManagement'
 import { useCountryIso } from '@client/hooks'
 import { useToaster } from '@client/hooks/useToaster'
