+++ src/client/store/ui/userManagement/actions/updateSectionAuth.ts
